chore: split cmds
