+++ .env.example
@@ -1,1 +1,2 @@
 RUST_LOG=debug
+IMAGE_MAGICK_DIR=
